agents/ai-consultant-agent/requirements.txt
google-adk>=0.0.2
google-genai>=1.5.0,<2.0.0
pydantic>=2.10.6,<3.0.0
python-dotenv>=1.0.1
requests>=2.31.0
google-cloud-aiplatform[adk,agent-engines]>=1.88.0
absl-py>=2.2.1,<3.0.0
itellm>=1.0.0